REFACTOR: Modificación signup.py
